Changed up processing of comands.
Now the strings that are selected in the command window are checked against strings that were loaded into a map to make sure if another language is being used that the strings do in fact match. Also ParseChattingCommand has been changed to  ParseCommand since we're using it for more than just parsing the chat's commands.